+++ shared/lib/types.ts
@@ -74,6 +74,9 @@ interface ChannelConfigurationOptions {
     }>
     // TODO: bannedJIDs: string[]
   }
+  slowMode: {
+    defaultDelay: number
+  }
 }
 
 interface ChannelConfiguration {