+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Slow mode (#192):
   * new option in room configuration to set the slow mode delay (new prosody module mod_muc_slow_mode).
+  * default delay is configurable in channel's chat rooms options.
 
 ### Minor changes and fixes